feat: add validation for MongoDB ObjectId

+++ src/module/accommodation/accommodation.service.js
@@ -2,6 +2,7 @@ const autoBind = require("auto-bind");
 const createHttpError = require("http-errors");
 const { AccommodationModel } = require("./accommodation.model");
 const { AccommodationMessages } = require("./accommodation.messages");
+const { isValidObjectId } = require("mongoose");
 class AccommodationService {
   #model;
   constructor() {
@@ -17,6 +18,8 @@ class AccommodationService {
     return true;
   }
   async remove(id) {
+    if(!isValidObjectId(id))
+      throw new createHttpError.BadRequest('id معتبر نیست.')
     const accommodation = await this.#model.findOneAndDelete({ _id: id });
     if (!accommodation)
       throw new createHttpError.NotFound(AccommodationMessages.NotFound);
